+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.6.2] - 2026-04-09
+
+### Fixed
+
+- Fix JGrid blank color name on FlatLaf [#558](https://github.com/nbbrd/java-desktop-util/issues/558)
+
 ## [2.6.1] - 2026-02-19
 
 ### Changed
@@ -98,7 +104,7 @@ java-desktop-util follows [semantic versioning](https://semver.org/).
 
 ### Changed
 
-- Migration to [Maven-Central](https://search.maven.org/search?q=g:com.github.nbbrd.java-desktop-util)
+- Migration to [Maven-Central](https://search.maven.org)
 - **Breaking change:** Maven groupId is now `com.github.nbbrd.java-desktop-util`
 
 ## [2.2.4] - 2020-05-25
@@ -141,7 +147,8 @@ java-desktop-util follows [semantic versioning](https://semver.org/).
 
 - JDK11 cleanup
 
-[Unreleased]: https://github.com/nbbrd/java-desktop-util/compare/v2.6.1...HEAD
+[Unreleased]: https://github.com/nbbrd/java-desktop-util/compare/v2.6.2...HEAD
+[2.6.2]: https://github.com/nbbrd/java-desktop-util/compare/v2.6.1...v2.6.2
 [2.6.1]: https://github.com/nbbrd/java-desktop-util/compare/v2.6.0...v2.6.1
 [2.6.0]: https://github.com/nbbrd/java-desktop-util/compare/v2.5.0...v2.6.0
 [2.5.0]: https://github.com/nbbrd/java-desktop-util/compare/v2.4.0...v2.5.0